more agressive cleaning up
